src/loading/requirements.txt
kafka-python
pymongo
psycopg2-binary
redis
prometheus-client
pytest